requirements.txt
python-telegram-bot==20.0
requests==2.25.1
python-dotenv==0.19.1
gunicorn